requirements.txt
langchain==0.0.118
openai==0.27.2
python-dotenv==1.0.0
rich==13.3.2
transformers==4.27.2
keyboard~=0.13.5
tiktoken==0.3.2
discord==2.3.2